refactor(substrate): extract shared copyTree helper
substrate-adopt and substrate-copy each shipped a near-identical
recursive copy-with-chmod-preservation walk. Hoist a single copyTree
into substrate-paths with an optional exclude predicate; adopt passes
its exclude filter, copy passes none.
